flask/requirements.txt
flask=2.0.1
Flask-Cors
PyJWT
PyMySQL